survey_backend/requirements.txt
Django==4.0.4
djangorestframework==3.14.0
django-cors-headers==3.13.0
pytz==2022.1
sqlparse==0.4.2
asgiref==3.5.2
gunicorn==20.1.0
psycopg2==2.9.3
